u-boot: switch to upstream SRCREV for rk-u-boot-env
When the rk-u-boot-env feature was implemented, in order for it to work the
build needed a version of U-Boot that was slightly newer than 2024.04 (the
current release supported by Yocto at that time). Now that 2024.10 is out,
remove the SRCREV pin since the necessary code is now part of this release.

Run tested both with and without RK_RAUC_DEMO on:
	nanopi-m4-2gb
	nanopi-m4b
	nanopi-r2s
	rock-3a
	rock-5a
	rock-5b
	rock-pi-4b
	rock-pi-e
	rock-pi-s
	rock64

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>