rk3288: use u-boot.img as bootloader image
Currently, u-boot.bin is used for supported rk3288-based boards. But on
some boards, such as tinker-board, it can not be loaded, since
u-boot.bin is raw and those boards have raw image support
(CONFIG_SPL_RAW_IMAGE_SUPPORT) disabled in favour of FIT images.

Use u-boot.img which is FIT when FIT support is enabled, and raw
otherwise.

Also add a note that u-boot.itb should be used if one wantes to utilize
OP-TEE.

Suggested-by: Quentin Schulz <quentin.schulz@cherry.de>

NOTE: a small change is needed to rockchip-gpt-img.bbclass so
      rock2-square will continue to build, this was done and
      rolled into the commit before pushing

Tweaked-by: Trevor Woerner <twoerner@gmail.com>
Reviewed-by: Quentin Schulz <quentin.schulz@cherry.de>
Signed-off-by: Peter Shkenev <mustela@erminea.space>