abd-partition: place unit file in system location
The system location for systemd unit files is in /usr/lib/systemd and
local or admin-created unit files, that are meant to take precedence
over the units in the system location, are to be placed in /etc/systemd.
Move these unit files to the system location instead of putting them in
the override location.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>